+++ packages/react-tabs/CHANGELOG.md
@@ -1,8 +1,45 @@
 # Change Log - @fluentui/react-tabs
 
-This log was last generated on Tue, 13 Jul 2021 07:35:36 GMT and should not be manually modified.
+This log was last generated on Fri, 16 Jul 2021 00:35:31 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [1.0.0-beta.127](https://github.com/microsoft/fluentui/tree/@fluentui/react-tabs_v1.0.0-beta.127)
+
+Fri, 16 Jul 2021 00:35:31 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-tabs_v1.0.0-beta.126..@fluentui/react-tabs_v1.0.0-beta.127)
+
+### Patches
+
+- Bump @fluentui/react to v8.23.4 ([PR #18909](https://github.com/microsoft/fluentui/pull/18909) by ilrosen@microsoft.com)
+
+## [1.0.0-beta.126](https://github.com/microsoft/fluentui/tree/@fluentui/react-tabs_v1.0.0-beta.126)
+
+Thu, 15 Jul 2021 07:36:18 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-tabs_v1.0.0-beta.125..@fluentui/react-tabs_v1.0.0-beta.126)
+
+### Patches
+
+- Bump @fluentui/react to v8.23.3 ([PR #18884](https://github.com/microsoft/fluentui/pull/18884) by behowell@microsoft.com)
+
+## [1.0.0-beta.125](https://github.com/microsoft/fluentui/tree/@fluentui/react-tabs_v1.0.0-beta.125)
+
+Wed, 14 Jul 2021 07:28:19 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-tabs_v1.0.0-beta.124..@fluentui/react-tabs_v1.0.0-beta.125)
+
+### Patches
+
+- Bump @fluentui/react to v8.23.2 ([PR #18853](https://github.com/microsoft/fluentui/pull/18853) by makopch@microsoft.com)
+
+## [1.0.0-beta.124](https://github.com/microsoft/fluentui/tree/@fluentui/react-tabs_v1.0.0-beta.124)
+
+Tue, 13 Jul 2021 22:32:58 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-tabs_v1.0.0-beta.123..@fluentui/react-tabs_v1.0.0-beta.124)
+
+### Patches
+
+- Bump @fluentui/react to v8.23.1 ([PR #18890](https://github.com/microsoft/fluentui/pull/18890) by behowell@microsoft.com)
+- Bump @fluentui/react-conformance to v0.4.3 ([PR #18925](https://github.com/microsoft/fluentui/pull/18925) by elcraig@microsoft.com)
 
 ## [1.0.0-beta.123](https://github.com/microsoft/fluentui/tree/@fluentui/react-tabs_v1.0.0-beta.123)